core: rename connection_manager to transport_manager
Rename connection_manager module and related fields/methods to transport_manager, updating imports, exports, tests, and ModelClient/Session plumbing to use TransportManager consistently.
