+++ codex-rs/core/src/transport_manager.rs
@@ -4,10 +4,6 @@ use std::sync::atomic::Ordering;
 
 use crate::model_provider_info::WireApi;
 
-/// Session-scoped connection transport state.
-///
-/// Once fallback is activated, the session will stick to HTTP for the rest of
-/// its lifetime when the provider prefers the WebSocket transport.
 #[derive(Clone, Debug, Default)]
 pub struct TransportManager {
     fallback_to_http: Arc<AtomicBool>,
@@ -28,8 +24,6 @@ impl TransportManager {
         }
     }
 
-    /// Activates sticky HTTP fallback. Returns `true` if this call flipped the
-    /// state from WebSocket to HTTP.
     pub fn activate_http_fallback(&self, provider_wire_api: WireApi) -> bool {
         provider_wire_api == WireApi::ResponsesWebsocket
             && !self.fallback_to_http.swap(true, Ordering::Relaxed)
