+++ codex-rs/core/tests/chat_completions_sse.rs
@@ -4,13 +4,13 @@ use std::sync::Arc;
 use tracing_test::traced_test;
 
 use codex_core::CodexAuth;
-use codex_core::TransportManager;
 use codex_core::ContentItem;
 use codex_core::ModelClient;
 use codex_core::ModelProviderInfo;
 use codex_core::Prompt;
 use codex_core::ResponseEvent;
 use codex_core::ResponseItem;
+use codex_core::TransportManager;
 use codex_core::WireApi;
 use codex_core::models_manager::manager::ModelsManager;
 use codex_otel::OtelManager;
